+++ page-hunter/src/page_hunter/errors.rs
@@ -3,7 +3,7 @@ use std::fmt::{Debug, Display, Formatter, Result};
 #[allow(unused_imports)]
 use super::models::Page;
 
-#[cfg(any(feature = "pg-sqlx"))]
+#[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
 use sqlx::Error as SqlxError;
 
 /// Provides a way to categorize the pagination error.
@@ -12,8 +12,12 @@ pub enum ErrorKind {
     FieldValueError(String),
 
     /// Raised when an error occurs during a database operation.
-    #[cfg(any(feature = "pg-sqlx"))]
+    #[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
     DatabaseError(String),
+
+    /// Raised when an error occurs while trying to convert a [`sqlx::Row`] into a struct `S` using [`sqlx`] features.
+    #[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
+    FromRowError(String),
 }
 
 impl ErrorKind {
@@ -22,10 +26,16 @@ impl ErrorKind {
         matches!(self, ErrorKind::FieldValueError(_))
     }
 
-    #[cfg(any(feature = "pg-sqlx"))]
     /// Check if the [`ErrorKind`] is a [`ErrorKind::DatabaseError`].
+    #[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
     pub fn is_database_error(&self) -> bool {
         matches!(self, ErrorKind::DatabaseError(_))
+    }
+
+    /// Check if the [`ErrorKind`] is a [`ErrorKind::FromRowError`].
+    #[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
+    pub fn is_from_row_error(&self) -> bool {
+        matches!(self, ErrorKind::FromRowError(_))
     }
 }
 
@@ -35,8 +45,11 @@ impl Display for ErrorKind {
         match self {
             ErrorKind::FieldValueError(detail) => write!(f, "FIELD VALUE ERROR- {}", detail),
 
-            #[cfg(any(feature = "pg-sqlx"))]
+            #[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
             ErrorKind::DatabaseError(detail) => write!(f, "DATABASE ERROR- {}", detail),
+
+            #[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
+            ErrorKind::FromRowError(detail) => write!(f, "FROM ROW ERROR- {}", detail),
         }
     }
 }
@@ -47,8 +60,11 @@ impl Debug for ErrorKind {
         match self {
             ErrorKind::FieldValueError(detail) => write!(f, "FieldValueError({:?})", detail),
 
-            #[cfg(any(feature = "pg-sqlx"))]
+            #[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
             ErrorKind::DatabaseError(detail) => write!(f, "DatabaseError({:?})", detail),
+
+            #[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
+            ErrorKind::FromRowError(detail) => write!(f, "FromRowError({:?})", detail),
         }
     }
 }
@@ -59,8 +75,11 @@ impl Clone for ErrorKind {
         match self {
             ErrorKind::FieldValueError(detail) => ErrorKind::FieldValueError(detail.to_owned()),
 
-            #[cfg(any(feature = "pg-sqlx"))]
+            #[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
             ErrorKind::DatabaseError(detail) => ErrorKind::DatabaseError(detail.to_owned()),
+
+            #[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
+            ErrorKind::FromRowError(detail) => ErrorKind::FromRowError(detail.to_owned()),
         }
     }
 }
@@ -108,7 +127,7 @@ impl From<ErrorKind> for PaginationError {
 }
 
 /// Implementation of [`From<SqlxError>`] for [`PaginationError`].
-#[cfg(any(feature = "pg-sqlx"))]
+#[cfg(any(feature = "pg-sqlx", feature = "mysql-sqlx"))]
 impl From<SqlxError> for PaginationError {
     fn from(value: SqlxError) -> Self {
         Self {