+++ page-hunter/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added:
 
+- 🧑🏻‍💻 Implement `ErrorKind::FromRowError`.
 - 🧑🏻‍💻 Implement DB migrations with sqlx for the creation of the MySQL test database.
 - 🧑🏻‍💻 Implement ***SqlxPagination*** to generalize the implementation of pagination methods using *sqlx*.
 - 🧑🏻‍💻 Implement **SqlxPagination** for `QueryBuilder<MySQL>` to paginate results from a SQL query into a `Page`. Only available on ***mysql-sqlx*** feature is enabled.